Added 16/32 bit address flag to file format and check for it.
MinorVersion is now 1 byte and the next byte is a 0 if this file has 16 bit addresses or 1 if they are 32 bits. Check in Decompiler and Interpreter to make sure address size is correct for how they were compiled

+++ CodeGen.h
@@ -31,7 +31,7 @@ class CodeGen
     virtual ~CodeGen() { }
     
     virtual uint16_t majorVersion() const = 0;
-    virtual uint16_t minorVersion() const = 0;
+    virtual uint8_t minorVersion() const = 0;
     
     virtual void emitCode(const ASTPtr& node, bool isLHS) = 0;
