+++ CodeGenStackVM.h
@@ -19,7 +19,7 @@ class CodeGenStackVM : public CodeGen
     virtual ~CodeGenStackVM() { }
     
     virtual uint16_t majorVersion() const override { return 1; }
-    virtual uint16_t minorVersion() const override { return 0; }
+    virtual uint8_t minorVersion() const override { return 0; }
 
     virtual void emitCode(const ASTPtr& node, bool isLHS) override;
